# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bsllmner-viewer API",
    "description": "Queries, aggregations, and exports over ontology-mapped BioSample annotations. The conventions follow the DDBJ Search API.\n\nStart with [/llms.txt](/llms.txt): condition examples and recipes for common tasks. [/llms-full.txt](/llms-full.txt) has the rules that span operations: the condition language, the descendants of terms, counting units and populations, self-exclusion, default elements, expected counts, export formats, errors, limits, and caching with `ETag` and `If-None-Match`. A description in this document names the heading of the rule that it relies on.",
    "contact": {
      "name": "BioData Science Initiative",
      "url": "https://github.com/dbcls/bsllmner-viewer"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/export/accessions/{type}": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export the accessions of the entries that match a condition",
        "description": "Returns every distinct accession of `type` among the entries that match `q`, as plain text: a header line that starts with `#` and names `q` and the dataset version, then one accession per line in ascending order. See \"Entries\" in /llms-full.txt.",
        "operationId": "exportAccessions",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AccessionType",
              "description": "Kind of accession to list"
            },
            "description": "Kind of accession to list"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt",
              "examples": [
                "disease:\"MONDO:0007254\"",
                "cell_line:\"CVCL:0030\" AND library_strategy:ATAC-seq",
                "organism_id:9606 AND date_published:[2015-01-01 TO 2020-12-31]",
                "disease_status:unmapped",
                "\"breast cancer\" AND NOT organism_id:10090"
              ],
              "title": "Q"
            },
            "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt"
          }
        ],
        "responses": {
          "200": {
            "description": "The header line and the accessions",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "headers": {
              "X-Dataset-Version": {
                "description": "Dataset version of the export: the name as a JSON string, the creation time, and the digest",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (the request does not match the operation)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. The slug that ends `type` is one of:\n\n- `unexpected-token`: The condition does not parse, is longer than 4096 characters, or is blank where a condition is required, or a keyword has a wildcard. Parse, select, and keyword also return this slug if the condition that they return would be longer than 4096 characters. Fix the condition at the column that `detail` gives. Quote a value that has a colon or a space. See \"Grammar\" in /llms-full.txt.\n- `unknown-field`: A condition, a clause, or a parameter (`field`, `row`, `col`) names a field that the dataset does not have. Use a field that `dslFields` of `GET /api/dataset` lists.\n- `invalid-date-format`: A value of `date_published` in the form `YYYY-MM-DD` is not a calendar date, such as `2020-13-45`, or a bound of a range is not a date in that form. Write the date as `YYYY-MM-DD`, such as `2020-12-31`. See \"Fields\" in /llms-full.txt.\n- `invalid-operator-for-field`: The form of a value does not fit the field, such as a range or a date on a field other than `date_published`, a value with a wildcard, or a value of `date_published` that is neither a date in the form `YYYY-MM-DD` nor a range (`2020-1-1`, a quoted date, or `\"\"`). Use `field:value` for the field, `date_published:YYYY-MM-DD` for a date, or `date_published:[a TO b]` for a range. See \"Fields\" in /llms-full.txt.\n- `invalid-value`: A value is not allowed for the field: a status that is not a status group, an assay that the dataset does not have, a term value that is not a term ID, an organism that is not a number, a keyword without a letter or a digit, or keywords over their limits. Use the values that `detail` lists. Find a term ID with `GET /api/terms`. See \"Fields\" and \"Keywords\" in /llms-full.txt.\n- `nest-depth-exceeded`: Groups nest deeper than 5 levels, or the condition has more than 512 nodes. Flatten the groups or split the condition. See \"Limits\" in /llms-full.txt.\n- `missing-value`: A clause on a field other than `date_published` has an empty value. Give the clause a value. See \"Fields\" in /llms-full.txt.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable. The slug that ends `type` is one of:\n\n- `server-busy`: The api has no free slot for the request after waiting, or too many requests wait already. Wait for the `Retry-After` seconds and send the request again.\n- `query-timeout`: The query took longer than the time limit of a request and was stopped. Narrow the condition. The same request does not finish by itself.\n- `query-too-large`: The query needs more memory or temporary disk space than a worker may use and was stopped. Narrow the condition, or ask for fewer elements.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/export/entries/{type}": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export the BioSamples that match a condition as TSV or NDJSON",
        "description": "Returns every BioSample that matches `q`, in the order of the entry list, as TSV or as NDJSON. NDJSON is newline-delimited JSON: one JSON object per line. See \"Entries\" in /llms-full.txt for the columns and the cells.",
        "operationId": "exportEntries",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntryType",
              "description": "Entry type. Only `biosample` has entries"
            },
            "description": "Entry type. Only `biosample` has entries"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt",
              "examples": [
                "disease:\"MONDO:0007254\"",
                "cell_line:\"CVCL:0030\" AND library_strategy:ATAC-seq",
                "organism_id:9606 AND date_published:[2015-01-01 TO 2020-12-31]",
                "disease_status:unmapped",
                "\"breast cancer\" AND NOT organism_id:10090"
              ],
              "title": "Q"
            },
            "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "tsv",
                "ndjson"
              ],
              "type": "string",
              "description": "File format of the entries",
              "default": "tsv",
              "title": "Format"
            },
            "description": "File format of the entries"
          }
        ],
        "responses": {
          "200": {
            "description": "The entries, one per line after the header line of the TSV",
            "content": {
              "text/tab-separated-values": {
                "schema": {
                  "type": "string"
                }
              },
              "application/x-ndjson": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "headers": {
              "X-Dataset-Version": {
                "description": "Dataset version of the export: the name as a JSON string, the creation time, and the digest",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (the request does not match the operation)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. The slug that ends `type` is one of:\n\n- `unexpected-token`: The condition does not parse, is longer than 4096 characters, or is blank where a condition is required, or a keyword has a wildcard. Parse, select, and keyword also return this slug if the condition that they return would be longer than 4096 characters. Fix the condition at the column that `detail` gives. Quote a value that has a colon or a space. See \"Grammar\" in /llms-full.txt.\n- `unknown-field`: A condition, a clause, or a parameter (`field`, `row`, `col`) names a field that the dataset does not have. Use a field that `dslFields` of `GET /api/dataset` lists.\n- `invalid-date-format`: A value of `date_published` in the form `YYYY-MM-DD` is not a calendar date, such as `2020-13-45`, or a bound of a range is not a date in that form. Write the date as `YYYY-MM-DD`, such as `2020-12-31`. See \"Fields\" in /llms-full.txt.\n- `invalid-operator-for-field`: The form of a value does not fit the field, such as a range or a date on a field other than `date_published`, a value with a wildcard, or a value of `date_published` that is neither a date in the form `YYYY-MM-DD` nor a range (`2020-1-1`, a quoted date, or `\"\"`). Use `field:value` for the field, `date_published:YYYY-MM-DD` for a date, or `date_published:[a TO b]` for a range. See \"Fields\" in /llms-full.txt.\n- `invalid-value`: A value is not allowed for the field: a status that is not a status group, an assay that the dataset does not have, a term value that is not a term ID, an organism that is not a number, a keyword without a letter or a digit, or keywords over their limits. Use the values that `detail` lists. Find a term ID with `GET /api/terms`. See \"Fields\" and \"Keywords\" in /llms-full.txt.\n- `nest-depth-exceeded`: Groups nest deeper than 5 levels, or the condition has more than 512 nodes. Flatten the groups or split the condition. See \"Limits\" in /llms-full.txt.\n- `missing-value`: A clause on a field other than `date_published` has an empty value. Give the clause a value. See \"Fields\" in /llms-full.txt.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable. The slug that ends `type` is one of:\n\n- `server-busy`: The api has no free slot for the request after waiting, or too many requests wait already. Wait for the `Retry-After` seconds and send the request again.\n- `query-timeout`: The query took longer than the time limit of a request and was stopped. Narrow the condition. The same request does not finish by itself.\n- `query-too-large`: The query needs more memory or temporary disk space than a worker may use and was stopped. Narrow the condition, or ask for fewer elements.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessionType": {
        "type": "string",
        "enum": [
          "biosample",
          "sra-experiment",
          "sra-run",
          "bioproject"
        ]
      },
...
      "EntryType": {
        "type": "string",
        "const": "biosample"
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "`about:blank` for an error that the HTTP status describes. Otherwise a URI that ends with the slug of the rule that the request broke. The description of the 400 response of each operation lists its slugs."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "The HTTP status phrase"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance",
            "description": "The path of the request"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 time in UTC"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid",
            "description": "The value of the `X-Request-ID` header of the response"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "timestamp",
          "requestId"
        ],
        "title": "ProblemDetails",
        "description": "RFC 7807 problem details of an error response."
      },
...
    }
  },
...
}